Add custom tree content renderers
